requirements.txt
slack-sdk
python-dotenv
flask
black
requests
semantic_version